requirements.txt
telegram~=0.0.1
PyYAML~=5.3.1
bs4~=0.0.1
beautifulsoup4~=4.8.2
requests~=2.22.0
todoist~=0.0.1
todoist-python~=8.1.3
google-api-python-client~=2.10.0